refactor: new gdn action selection file
